sipdroid: redo codec infrastructure to support arbitrary codecs
Instead of having the GSM codec as special, this patch uses a plugin
infrastructure to support the adding of arbitrary codecs to Siproid.
It also changes the way users select codecs by adding a new 'Audio
Codecs' preference screen and allowing them to select for each codec
when it will be used (always, edge only or never) and also allowing
them to change the order in which codecs are presented to the other
end.

This code also fixes a bug in the original implementation where on an
incoming call, we could be presented with a codec we support that
actually has a different media identifier number in the session
descriptor.  On incoming, it is the remote who chooses the media
identifier numbers to use, so we have to do a string comparison to see
if we support it and then globally use the remote's media identifier
number.

I also added the codec being used to the in call screen and a long
delay call disconnect dialog if we find a codec mismatch on either
incoming or outgoing.

Signed-off-by: James Bottomley <James.Bottomley@HansenPartnership.com>

git-svn-id: http://sipdroid.googlecode.com/svn/trunk@459 8b64e154-0f13-11de-9106-f38757f67524