Some UI changes required for Shiny release
Removed Camera permission as unnecessary in Manifest
Amended Maps message to redirect to Wiki and be more positive in the
potential for release of this app.
Maps help currently commented out. The help screens are in a state of
flux as major improvements are being done for release.
DTMF and Speaker buttons commented out as unsupported functions.