Changed Help System to local HTML files.
The Help Subsystem of UI is changed to local webview. The idea is to
eventually have all help files written once on Wiki, then placed as
local HTML files for Android, rather than having to convert to strings
etc. NOTE: refactoring to remove unwanted strings, icons, remnants of
older software is a very good project for an undergrad.