build_setup.sh: do not run libsodium's autogen.sh
Do not invoke libsodium's autogen.sh script, as all the Makefile.in
files etc. are already included in the libsodium repository.  Running
autogen.sh simply creates local modifications in the libsodium working
copy (from a newer version of autotools) that are unnecessary and
interfere with development work.

Fix copyright message.